--- a/cover.h
+++ b/cover.h
@@ -13,14 +13,8 @@
 #  define ZDICT_STATIC_LINKING_ONLY
 #endif
 
-#include <stdio.h>  /* fprintf */
-#include <stdlib.h> /* malloc, free, qsort */
-#include <string.h> /* memset */
-#include <time.h>   /* clock */
-#include "mem.h" /* read */
-#include "pool.h"
-#include "threading.h"
-#include "zstd_internal.h" /* includes zstd.h */
+#include "threading.h" /* ZSTD_pthread_mutex_t */
+#include "mem.h"   /* U32, BYTE */
 #include "zdict.h"
 
 /**