Same workflow as https://github.com/DataDog/zstd/pull/125

- [x] Run below:
```
git clone https://github.com/facebook/zstd.git c
cd c
git checkout v1.5.6
cd ..
git clone git@github.com:DataDog/zstd.git go
find c/lib \( -name '*.h' -o -name '*.c' -o -name '*.S' \) -exec cp -v {} go/ \;
```

- [x] Run `python tools/flatten_imports.py`
- [x] Run `python tools/insert_libzstd_ifdefs.py`


This is the code as of https://github.com/facebook/zstd/releases/tag/v1.5.6 (https://github.com/facebook/zstd/commit/794ea1b0afca0f020f4e57b6732332231fb23c70)

Benchmarks will come in the next comment